Rename main view package
