+++ spotify-clone-frontend/app/src/main/java/com/spotifyclone/presentation/main/ReplaceListener.kt
@@ -1,0 +1,4 @@
+package com.spotifyclone.presentation.main
+
+interface ReplaceListener {
+}
