--- spotify-clone-frontend/app/src/main/java/com/spotifyclone/presentation/maintab/ReplaceListener.kt
@@ -1,7 +1,0 @@
-package com.spotifyclone.presentation.maintab
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-
-interface ReplaceListener {
-}
