+++ spotify-clone-frontend/app/src/main/res/layout/fragment_music_player.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="80dp"
+    android:layout_above="@id/bottomNavMenu"
     android:orientation="vertical">
 
     <ProgressBar
